fix: Fix the bug on download zk binary file.
